Remove named returns

+++ contracts/base/SafeMath.sol
@@ -1,23 +1,23 @@
 pragma solidity 0.4.11;
 
 contract SafeMath {
-    function safeMul(uint a, uint b) internal constant returns (uint) {
+    function safeMul(uint a, uint b) internal constant returns (uint256) {
         uint c = a * b;
         require(a == 0 || c / a == b);
         return c;
     }
 
-    function safeDiv(uint a, uint b) internal constant returns (uint) {
+    function safeDiv(uint a, uint b) internal constant returns (uint256) {
         uint c = a / b;
         return c;
     }
 
-    function safeSub(uint a, uint b) internal constant returns (uint) {
+    function safeSub(uint a, uint b) internal constant returns (uint256) {
         require(b <= a);
         return a - b;
     }
 
-    function safeAdd(uint a, uint b) internal constant returns (uint) {
+    function safeAdd(uint a, uint b) internal constant returns (uint256) {
         uint c = a + b;
         require(c >= a);
         return c;
